Fixing problems building on linux, where size_t and uint64_t are the same. Simplified some hash() implementations by removing the hashing of size - hashing of the children should be sufficient. Note that when hashing a size_t (as in the CurvesPrimitive::hash() implementation), you should cast to uint64_t - this should keep the hash identical whether created on a 32 or 64 bit machine.

+++ src/IECore/PointsPrimitive.cpp
@@ -109,7 +109,7 @@ void PointsPrimitive::memoryUsage( Object::MemoryAccumulator &a ) const
 void PointsPrimitive::hash( MurmurHash &h ) const
 {
 	Primitive::hash( h );
-	h.append( m_numPoints );
+	h.append( (uint64_t)m_numPoints );
 }
 
 size_t PointsPrimitive::getNumPoints() const
